+++ packages/utils/functions/token/index.ts
@@ -1,4 +1,5 @@
 export * from './decodeJWT'
+export * from './getAccessToken'
 export * from './getToken'
 export * from './isUserTokenValid'
 export * from './refreshAccessToken'